requirements.txt
flask
opencv-python-headless
google-cloud-storage
requests
